Removed redundant code, and changed get_events
